+++ python_installer.sh
@@ -3,4 +3,9 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 echo 'Create symbolic links for .pyrc'
-ln -s $DIR/.pyrc ~/.pyrc
+if [ ! -f ~/.pyrc ]
+then
+  ln -s $DIR/.pyrc ~/.pyrc
+else
+  echo '  ~/.pyrc already exists'
+fi
