+++ vim_installer.sh
@@ -17,10 +17,34 @@ pause
 
 
 echo 'Create symbolic links for tmux'
-ln -s $DIR/.tmux.conf ~/.tmux.conf
+if [ ! -f ~/.tmux.conf ]
+then
+  ln -s $DIR/.tmux.conf ~/.tmux.conf
+else
+  echo '  ~/.tmux.conf already exists'
+fi
+
 echo 'Create symbolic links for vimrc & plugins'
-ln -s $DIR/.vimrc ~/.vimrc
-ln -s $DIR/vim/rcfiles ~/.vim/rcfiles
-ln -s $DIR/vim/rcplugins ~/.vim/rcplugins
+if [ ! -f ~/.vimrc ]
+then
+  ln -s $DIR/.vimrc ~/.vimrc
+else
+  echo '  ~/.vimrc already exists'
+fi
+
+if [ ! -d ~/.vim/rcfiles ]
+then
+  ln -s $DIR/vim/rcfiles ~/.vim/rcfiles
+else
+  echo '  ~/.vim/rcfiles already exists'
+fi
+
+if [ ! -d ~/.vim/rcplugins ]
+then
+  ln -s $DIR/vim/rcplugins ~/.vim/rcplugins
+else
+  echo '  ~/.vim/rcplugins already exists'
+fi
+
 # ln -s $DIR/vim/snippets ~/.vim/snippets
 # ln -s $DIR/bin/* ~/.bin/*
